+++ PredicateFlow-Realm/Classes/LinkingObj+PF.swift
@@ -8,11 +8,11 @@
 import RealmSwift
 
 extension LinkingObjects {
-    
+
     public func filter(_ predicateResult: PredicateResult) -> Results<Element> {
         return filter(predicateResult.query())
     }
-    
+
     public func sorted(_ sortDescriptions: SortDescriptor...) -> Results<Element> {
         return sorted(by: sortDescriptions)
     }