+++ PredicateFlow/Classes/Utils/GeneratedPredicateSchema.swift
@@ -12,23 +12,23 @@ Used for defining generated predicate schema.
 */
 public protocol GeneratedPredicateSchema {
 
-	/**
-	Init with a prev compound field builder.
-	*/
-	init(compoundFieldBuilder: CompoundFieldBuilder)
+    /// Use the builder property if you need to build a predicate property
+    var compoundFieldBuilder: CompoundFieldBuilder { get }
 
-	/**
-	The entity name of the related schema. (e.g: if you define a class/struct Dog, this library will generate a related DogSchema that implements this protocol).
-	*/
-	static var entityName: String { get }
+	/// Init with a prev compound field builder.
+	init(compoundFieldBuilder: CompoundFieldBuilder)
 }
 
 public extension GeneratedPredicateSchema {
 
-	/**
-	Init with an empty compound field builder.
-	*/
+    /// Init with an empty compound field builder.
 	public init() {
 		self.init(compoundFieldBuilder: CompoundFieldBuilder())
 	}
+
+    /// A builder to create a new predicate property.
+    /// If the code generator does not generate correctly a property, you can use this builder to create a new predicate property.
+    public var builder: PredicatePropertyBuilder {
+        return PFPredicatePropertyBuilder(compoundFieldBuilder)
+    }
 }